+++ src/Builder/FormBuilder.php
@@ -113,29 +113,17 @@ class FormBuilder
 
 		$inputs = $this->getAccountSettingsSection($isApiKeyProvided);
 
-		$fields = [
-			'form' => [
-				'input' => $inputs,
-				'submit' => [
-					'title' => $this->module->l('Save'),
-					'class' => 'btn btn-default pull-right',
-				],
-			],
-		];
+        $helper = new HelperForm();
 
-		$helper = new HelperForm();
+        $helper->show_toolbar = false;
+        $helper->table = $this->module->getTable();
+        $helper->module = $this->module;
+        $helper->default_form_language = $this->module->getContext()->language->id;
+        $helper->allow_employee_form_lang = Configuration::get('PS_BO_ALLOW_EMPLOYEE_FORM_LANG', 0);
 
-		$helper->show_toolbar = false;
-		$helper->table = $this->module->getTable();
-		$helper->module = $this->module;
-		$helper->default_form_language = $this->module->getContext()->language->id;
-		$helper->allow_employee_form_lang = Configuration::get('PS_BO_ALLOW_EMPLOYEE_FORM_LANG', 0);
-
-		$helper->identifier = $this->module->getIdentifier();
-		$helper->submit_action = 'submitmollie';
-		$helper->currentIndex = $this->module->getContext()->link->getAdminLink('AdminModules', false)
-			. "&configure={$this->module->name}&tab_module={$this->module->tab}&module_name={$this->module->name}";
-		$helper->token = Tools::getAdminTokenLite('AdminModules');
+        $helper->identifier = $this->module->getIdentifier();
+        $helper->submit_action = 'submitGeneralSettingsConfiguration';
+        $helper->token = Tools::getAdminTokenLite(Mollie::ADMIN_MOLLIE_GENERAL_SETTINGS_CONTROLLER);
 
 		$helper->tpl_vars = [
 			'fields_value' => $this->configFieldService->getConfigFieldsValues(),
@@ -143,7 +131,17 @@ class FormBuilder
 			'id_language' => $this->module->getContext()->language->id,
 		];
 
-		return $helper->generateForm([$fields]);
+		return $helper->generateForm([
+            [
+                'form' => [
+                    'input' => $inputs,
+                    'submit' => [
+                        'title' => $this->module->l('Save'),
+                        'class' => 'btn btn-default pull-right',
+                    ],
+                ],
+            ]
+        ]);
 	}
 
 	protected function getAccountSettingsSection($isApiKeyProvided)
@@ -431,386 +429,387 @@ class FormBuilder
 		return $input;
 	}
 
-	public function buildAdvancedSettingsForm()
-	{
-		$advancedSettings = 'advanced_settings';
-		$input = [];
-		$orderStatuses = [];
-		$orderStatuses = array_merge($orderStatuses, OrderState::getOrderStates($this->lang->id));
-		$input[] = [
-			'type' => 'select',
-			'label' => $this->module->l('Push Locale to Payment Screen', self::FILE_NAME),
-			'tab' => $advancedSettings,
-			'desc' => TagsUtility::ppTags(
-				$this->module->l('When activated, Mollie will use your webshop\'s [1]Locale[/1]. If not, the browser\'s locale will be used.', self::FILE_NAME),
-				[$this->module->display($this->module->getPathUri(), 'views/templates/admin/locale_wiki.tpl')]
-			),
-			'name' => Config::MOLLIE_PAYMENTSCREEN_LOCALE,
-			'options' => [
-				'query' => [
-					[
-						'id' => Config::PAYMENTSCREEN_LOCALE_SEND_WEBSITE_LOCALE,
-						'name' => $this->module->l('Yes, push webshop Locale', self::FILE_NAME),
-					],
-					[
-						'id' => Config::PAYMENTSCREEN_LOCALE_BROWSER_LOCALE,
-						'name' => $this->module->l('No, use browser\'s Locale', self::FILE_NAME),
-					],
-				],
-				'id' => 'id',
-				'name' => 'name',
-			],
-		];
+	private function getAdvancedSettingsSection()
+    {
+        $advancedSettings = 'advanced_settings';
+        $input = [];
+        $orderStatuses = [];
+        $orderStatuses = array_merge($orderStatuses, OrderState::getOrderStates($this->lang->id));
 
-		if (Config::isVersion17()) {
-			$input[] = [
-				'type' => 'select',
-				'label' => $this->module->l('Send order confirmation email', self::FILE_NAME),
-				'tab' => $advancedSettings,
-				'name' => Config::MOLLIE_SEND_ORDER_CONFIRMATION,
-				'options' => [
-					'query' => [
-						[
-							'id' => Config::ORDER_CONF_MAIL_SEND_ON_CREATION,
-							'name' => $this->module->l('When Order is created', self::FILE_NAME),
-						],
-						[
-							'id' => Config::ORDER_CONF_MAIL_SEND_ON_PAID,
-							'name' => $this->module->l('When Order is Paid', self::FILE_NAME),
-						],
-						[
-							'id' => Config::ORDER_CONF_MAIL_SEND_ON_NEVER,
-							'name' => $this->module->l('Never', self::FILE_NAME),
-						],
-					],
-					'id' => 'id',
-					'name' => 'name',
-				],
-			];
+        $input[] = [
+            'type' => 'select',
+            'label' => $this->module->l('Push Locale to Payment Screen', self::FILE_NAME),
+            'tab' => $advancedSettings,
+            'desc' => TagsUtility::ppTags(
+                $this->module->l('When activated, Mollie will use your webshop\'s [1]Locale[/1]. If not, the browser\'s locale will be used.', self::FILE_NAME),
+                [$this->module->display($this->module->getPathUri(), 'views/templates/admin/locale_wiki.tpl')]
+            ),
+            'name' => Config::MOLLIE_PAYMENTSCREEN_LOCALE,
+            'options' => [
+                'query' => [
+                    [
+                        'id' => Config::PAYMENTSCREEN_LOCALE_SEND_WEBSITE_LOCALE,
+                        'name' => $this->module->l('Yes, push webshop Locale', self::FILE_NAME),
+                    ],
+                    [
+                        'id' => Config::PAYMENTSCREEN_LOCALE_BROWSER_LOCALE,
+                        'name' => $this->module->l('No, use browser\'s Locale', self::FILE_NAME),
+                    ],
+                ],
+                'id' => 'id',
+                'name' => 'name',
+            ],
+        ];
 
-			if (Module::isEnabled(Config::EMAIL_ALERTS_MODULE_NAME)) {
-				$input[] = [
-					'type' => 'select',
-					'label' => $this->module->l('Send new order email to merchant', self::FILE_NAME),
-					'tab' => $advancedSettings,
-					'name' => Config::MOLLIE_SEND_NEW_ORDER,
-					'desc' => TagsUtility::ppTags(
-						$this->module->l('Change when \'new_order\' email to merchant is sent (When using PrestaShop Mail Alerts module)', self::FILE_NAME),
-						[$this->module->display($this->module->getPathUri(), 'views/templates/admin/locale_wiki.tpl')]
-					),
-					'options' => [
-						'query' => [
-							[
-								'id' => Config::NEW_ORDER_MAIL_SEND_ON_CREATION,
-								'name' => $this->module->l('When Order is created', self::FILE_NAME),
-							],
-							[
-								'id' => Config::NEW_ORDER_MAIL_SEND_ON_PAID,
-								'name' => $this->module->l('When Order is Paid', self::FILE_NAME),
-							],
-							[
-								'id' => Config::NEW_ORDER_MAIL_SEND_ON_NEVER,
-								'name' => $this->module->l('Never', self::FILE_NAME),
-							],
-						],
-						'id' => 'id',
-						'name' => 'name',
-					],
-				];
-			}
-		}
-		$input[] = [
-			'type' => 'select',
-			'label' => $this->module->l('When to create the invoice?', self::FILE_NAME),
-			'desc' => $this->module->display($this->module->getPathUri(), 'views/templates/admin/invoice_description.tpl'),
-			'tab' => $advancedSettings,
-			'name' => Config::MOLLIE_KLARNA_INVOICE_ON,
-			'options' => [
-				'query' => [
-					[
-						'id' => Config::MOLLIE_STATUS_KLARNA_AUTHORIZED,
-						'name' => $this->module->l('Accepted', self::FILE_NAME),
-					],
-					[
-						'id' => Config::MOLLIE_STATUS_KLARNA_SHIPPED,
-						'name' => $this->module->l('Shipped', self::FILE_NAME),
-					],
-				],
-				'id' => 'id',
-				'name' => 'name',
-			],
-		];
+        if (Config::isVersion17()) {
+            $input[] = [
+                'type' => 'select',
+                'label' => $this->module->l('Send order confirmation email', self::FILE_NAME),
+                'tab' => $advancedSettings,
+                'name' => Config::MOLLIE_SEND_ORDER_CONFIRMATION,
+                'options' => [
+                    'query' => [
+                        [
+                            'id' => Config::ORDER_CONF_MAIL_SEND_ON_CREATION,
+                            'name' => $this->module->l('When Order is created', self::FILE_NAME),
+                        ],
+                        [
+                            'id' => Config::ORDER_CONF_MAIL_SEND_ON_PAID,
+                            'name' => $this->module->l('When Order is Paid', self::FILE_NAME),
+                        ],
+                        [
+                            'id' => Config::ORDER_CONF_MAIL_SEND_ON_NEVER,
+                            'name' => $this->module->l('Never', self::FILE_NAME),
+                        ],
+                    ],
+                    'id' => 'id',
+                    'name' => 'name',
+                ],
+            ];
 
-		$messageStatus = $this->module->l('Status for %s payments', self::FILE_NAME);
-		$descriptionStatus = $this->module->l('`%s` payments get status `%s`', self::FILE_NAME);
-		$messageMail = $this->module->l('Send mails when %s', self::FILE_NAME);
-		$descriptionMail = $this->module->l('Send mails when transaction status becomes %s?, self::FILE_NAME', self::FILE_NAME);
-		$allStatuses = OrderState::getOrderStates($this->lang->id);
-		$allStatusesWithSkipOption = array_merge([['id_order_state' => 0, 'name' => $this->module->l('Skip this status', self::FILE_NAME), 'color' => '#565656']], $allStatuses);
+            if (Module::isEnabled(Config::EMAIL_ALERTS_MODULE_NAME)) {
+                $input[] = [
+                    'type' => 'select',
+                    'label' => $this->module->l('Send new order email to merchant', self::FILE_NAME),
+                    'tab' => $advancedSettings,
+                    'name' => Config::MOLLIE_SEND_NEW_ORDER,
+                    'desc' => TagsUtility::ppTags(
+                        $this->module->l('Change when \'new_order\' email to merchant is sent (When using PrestaShop Mail Alerts module)', self::FILE_NAME),
+                        [$this->module->display($this->module->getPathUri(), 'views/templates/admin/locale_wiki.tpl')]
+                    ),
+                    'options' => [
+                        'query' => [
+                            [
+                                'id' => Config::NEW_ORDER_MAIL_SEND_ON_CREATION,
+                                'name' => $this->module->l('When Order is created', self::FILE_NAME),
+                            ],
+                            [
+                                'id' => Config::NEW_ORDER_MAIL_SEND_ON_PAID,
+                                'name' => $this->module->l('When Order is Paid', self::FILE_NAME),
+                            ],
+                            [
+                                'id' => Config::NEW_ORDER_MAIL_SEND_ON_NEVER,
+                                'name' => $this->module->l('Never', self::FILE_NAME),
+                            ],
+                        ],
+                        'id' => 'id',
+                        'name' => 'name',
+                    ],
+                ];
+            }
+        }
+        $input[] = [
+            'type' => 'select',
+            'label' => $this->module->l('When to create the invoice?', self::FILE_NAME),
+            'desc' => $this->module->display($this->module->getPathUri(), 'views/templates/admin/invoice_description.tpl'),
+            'tab' => $advancedSettings,
+            'name' => Config::MOLLIE_KLARNA_INVOICE_ON,
+            'options' => [
+                'query' => [
+                    [
+                        'id' => Config::MOLLIE_STATUS_KLARNA_AUTHORIZED,
+                        'name' => $this->module->l('Accepted', self::FILE_NAME),
+                    ],
+                    [
+                        'id' => Config::MOLLIE_STATUS_KLARNA_SHIPPED,
+                        'name' => $this->module->l('Shipped', self::FILE_NAME),
+                    ],
+                ],
+                'id' => 'id',
+                'name' => 'name',
+            ],
+        ];
 
-		$statuses = [];
-		foreach (Config::getStatuses() as $name => $val) {
-			if (PaymentStatus::STATUS_AUTHORIZED === $name) {
-				continue;
-			}
+        $messageStatus = $this->module->l('Status for %s payments', self::FILE_NAME);
+        $descriptionStatus = $this->module->l('`%s` payments get status `%s`', self::FILE_NAME);
+        $messageMail = $this->module->l('Send mails when %s', self::FILE_NAME);
+        $descriptionMail = $this->module->l('Send mails when transaction status becomes %s?, self::FILE_NAME', self::FILE_NAME);
+        $allStatuses = OrderState::getOrderStates($this->lang->id);
+        $allStatusesWithSkipOption = array_merge([['id_order_state' => 0, 'name' => $this->module->l('Skip this status', self::FILE_NAME), 'color' => '#565656']], $allStatuses);
 
-			$val = (int) $val;
-			if ($val) {
-				$orderStatus = new OrderState($val);
-				$statusName = $orderStatus->getFieldByLang('name', $this->lang->id);
-				$desc = Tools::strtolower(
-					sprintf(
-						$descriptionStatus,
-						$this->module->lang($name),
-						$statusName
-					)
-				);
-			} else {
-				$desc = sprintf($this->module->l('`%s` payments do not get a status', self::FILE_NAME), $this->module->lang($name));
-			}
-			$statuses[] = [
-				'name' => $name,
-				'key' => @constant('Mollie\Config\Config::MOLLIE_STATUS_' . Tools::strtoupper($name)),
-				'value' => $val,
-				'description' => $desc,
-				'message' => sprintf($messageStatus, $this->module->lang($name)),
-				'key_mail' => @constant('Mollie\Config\Config::MOLLIE_MAIL_WHEN_' . Tools::strtoupper($name)),
-				'value_mail' => Configuration::get('MOLLIE_MAIL_WHEN_' . Tools::strtoupper($name)),
-				'description_mail' => sprintf($descriptionMail, $this->module->lang($name)),
-				'message_mail' => sprintf($messageMail, $this->module->lang($name)),
-			];
-		}
-		$input[] = [
-			'type' => 'mollie-h2',
-			'name' => '',
-			'tab' => $advancedSettings,
-			'title' => $this->module->l('Order statuses', self::FILE_NAME),
-		];
+        $statuses = [];
+        foreach (Config::getStatuses() as $name => $val) {
+            if (PaymentStatus::STATUS_AUTHORIZED === $name) {
+                continue;
+            }
 
-		foreach (array_filter($statuses, function ($status) {
-			return in_array($status['name'], [
-				Config::MOLLIE_AWAITING_PAYMENT,
-				PaymentStatus::STATUS_PAID,
-				OrderStatus::STATUS_COMPLETED,
-				PaymentStatus::STATUS_AUTHORIZED,
-				PaymentStatus::STATUS_CANCELED,
-				PaymentStatus::STATUS_EXPIRED,
-				RefundStatus::STATUS_REFUNDED,
-				PaymentStatus::STATUS_OPEN,
-				Config::PARTIAL_REFUND_CODE,
-				OrderStatus::STATUS_SHIPPING,
-			]);
-		}) as $status) {
-			if (!in_array($status['name'], [Config::PARTIAL_REFUND_CODE])) {
-				$input[] = [
-					'type' => 'switch',
-					'label' => $status['message_mail'],
-					'tab' => $advancedSettings,
-					'name' => $status['key_mail'],
-					'is_bool' => true,
-					'values' => [
-						[
-							'id' => 'active_on',
-							'value' => true,
-							'label' => $this->module->l('Enabled', self::FILE_NAME),
-						],
-						[
-							'id' => 'active_off',
-							'value' => false,
-							'label' => $this->module->l('Disabled', self::FILE_NAME),
-						],
-					],
-				];
-			}
+            $val = (int) $val;
+            if ($val) {
+                $orderStatus = new OrderState($val);
+                $statusName = $orderStatus->getFieldByLang('name', $this->lang->id);
+                $desc = Tools::strtolower(
+                    sprintf(
+                        $descriptionStatus,
+                        $this->module->lang($name),
+                        $statusName
+                    )
+                );
+            } else {
+                $desc = sprintf($this->module->l('`%s` payments do not get a status', self::FILE_NAME), $this->module->lang($name));
+            }
+            $statuses[] = [
+                'name' => $name,
+                'key' => @constant('Mollie\Config\Config::MOLLIE_STATUS_' . Tools::strtoupper($name)),
+                'value' => $val,
+                'description' => $desc,
+                'message' => sprintf($messageStatus, $this->module->lang($name)),
+                'key_mail' => @constant('Mollie\Config\Config::MOLLIE_MAIL_WHEN_' . Tools::strtoupper($name)),
+                'value_mail' => Configuration::get('MOLLIE_MAIL_WHEN_' . Tools::strtoupper($name)),
+                'description_mail' => sprintf($descriptionMail, $this->module->lang($name)),
+                'message_mail' => sprintf($messageMail, $this->module->lang($name)),
+            ];
+        }
+        $input[] = [
+            'type' => 'mollie-h2',
+            'name' => '',
+            'tab' => $advancedSettings,
+            'title' => $this->module->l('Order statuses', self::FILE_NAME),
+        ];
 
-			$isStatusAwaiting = Config::MOLLIE_AWAITING_PAYMENT === $status['name'];
-			$input[] = [
-				'type' => 'select',
-				'label' => $status['message'],
-				'tab' => $advancedSettings,
-				'desc' => $status['description'],
-				'name' => $status['key'],
-				'options' => [
-					'query' => $isStatusAwaiting ? $allStatuses : $allStatusesWithSkipOption,
-					'id' => 'id_order_state',
-					'name' => 'name',
-				],
-			];
-		}
-		$input = array_merge($input, [
-			[
-				'type' => 'mollie-h2',
-				'name' => '',
-				'tab' => $advancedSettings,
-				'title' => $this->module->l('Visual Settings', self::FILE_NAME),
-			],
-			[
-				'type' => 'select',
-				'label' => $this->module->l('Images', self::FILE_NAME),
-				'tab' => $advancedSettings,
-				'desc' => $this->module->l('Show big, normal or no payment method logos on checkout.', self::FILE_NAME),
-				'name' => Config::MOLLIE_IMAGES,
-				'options' => [
-					'query' => [
-						[
-							'id' => Config::LOGOS_HIDE,
-							'name' => $this->module->l('hide', self::FILE_NAME),
-						],
-						[
-							'id' => Config::LOGOS_NORMAL,
-							'name' => $this->module->l('normal', self::FILE_NAME),
-						],
-						[
-							'id' => Config::LOGOS_BIG,
-							'name' => $this->module->l('big', self::FILE_NAME),
-						],
-					],
-					'id' => 'id',
-					'name' => 'name',
-				],
-			],
-			[
-				'type' => 'text',
-				'label' => $this->module->l('CSS file', self::FILE_NAME),
-				'tab' => $advancedSettings,
-				'desc' => TagsUtility::ppTags(
-					$this->module->l('Leave empty for default stylesheet. Should include file path when set. Hint: You can use [1]{BASE}[/1], [1]{THEME}[/1], [1]{CSS}[/1], [1]{MOBILE}[/1], [1]{MOBILE_CSS}[/1] and [1]{OVERRIDE}[/1] for easy folder mapping.', self::FILE_NAME),
-					[$this->module->display($this->module->getPathUri(), 'views/templates/front/kbd.tpl')]
-				),
-				'name' => Config::MOLLIE_CSS,
-				'class' => 'long-text',
-			],
-		]);
-		$input[] = [
-			'type' => 'mollie-carriers',
-			'label' => $this->module->l('Shipment information', self::FILE_NAME),
-			'tab' => $advancedSettings,
-			'name' => Config::MOLLIE_TRACKING_URLS,
-			'depends' => Config::MOLLIE_API,
-			'depends_value' => Config::MOLLIE_ORDERS_API,
-			'carriers' => $this->carrierInformationService->getAllCarriersInformation($this->lang->id),
-		];
-		$input[] = [
-			'type' => 'mollie-carrier-switch',
-			'label' => $this->module->l('Automatically ship on marked statuses', self::FILE_NAME),
-			'tab' => $advancedSettings,
-			'name' => Config::MOLLIE_AUTO_SHIP_MAIN,
-			'desc' => $this->module->l('Enabling this feature will automatically send shipment information when an order gets marked status.', self::FILE_NAME),
-			'is_bool' => true,
-			'values' => [
-				[
-					'id' => 'active_on',
-					'value' => true,
-					'label' => $this->module->l('Enabled', self::FILE_NAME),
-				],
-				[
-					'id' => 'active_off',
-					'value' => false,
-					'label' => $this->module->l('Disabled', self::FILE_NAME),
-				],
-			],
-			'depends' => Config::MOLLIE_API,
-			'depends_value' => Config::MOLLIE_ORDERS_API,
-		];
-		$input[] = [
-			'type' => 'checkbox',
-			'label' => $this->module->l('Automatically ship when one of these statuses is reached', self::FILE_NAME),
-			'tab' => $advancedSettings,
-			'desc' => $this->module->l('If an order reaches one of these statuses the module will automatically send shipment information', self::FILE_NAME),
-			'name' => Config::MOLLIE_AUTO_SHIP_STATUSES,
-			'multiple' => true,
-			'values' => [
-				'query' => $orderStatuses,
-				'id' => 'id_order_state',
-				'name' => 'name',
-			],
-			'expand' => (count($orderStatuses) > 10) ? [
-				'print_total' => count($orderStatuses),
-				'default' => 'show',
-				'show' => ['text' => $this->module->l('Show', self::FILE_NAME), 'icon' => 'plus-sign-alt'],
-				'hide' => ['text' => $this->module->l('Hide', self::FILE_NAME), 'icon' => 'minus-sign-alt'],
-			] : null,
-			'depends' => Config::MOLLIE_API,
-			'depends_value' => Config::MOLLIE_ORDERS_API,
-		];
-		$orderStatuses = [
-			[
-				'name' => $this->module->l('Disable this status', self::FILE_NAME),
-				'id_order_state' => '0',
-			],
-		];
-		$orderStatuses = array_merge($orderStatuses, OrderState::getOrderStates($this->lang->id));
-		$orderStatusesCount = count($orderStatuses);
-		for ($i = 0; $i < $orderStatusesCount; ++$i) {
-			$orderStatuses[$i]['name'] = $orderStatuses[$i]['id_order_state'] . ' - ' . $orderStatuses[$i]['name'];
-		}
+        foreach (array_filter($statuses, function ($status) {
+            return in_array($status['name'], [
+                Config::MOLLIE_AWAITING_PAYMENT,
+                PaymentStatus::STATUS_PAID,
+                OrderStatus::STATUS_COMPLETED,
+                PaymentStatus::STATUS_AUTHORIZED,
+                PaymentStatus::STATUS_CANCELED,
+                PaymentStatus::STATUS_EXPIRED,
+                RefundStatus::STATUS_REFUNDED,
+                PaymentStatus::STATUS_OPEN,
+                Config::PARTIAL_REFUND_CODE,
+                OrderStatus::STATUS_SHIPPING,
+            ]);
+        }) as $status) {
+            if (!in_array($status['name'], [Config::PARTIAL_REFUND_CODE])) {
+                $input[] = [
+                    'type' => 'switch',
+                    'label' => $status['message_mail'],
+                    'tab' => $advancedSettings,
+                    'name' => $status['key_mail'],
+                    'is_bool' => true,
+                    'values' => [
+                        [
+                            'id' => 'active_on',
+                            'value' => true,
+                            'label' => $this->module->l('Enabled', self::FILE_NAME),
+                        ],
+                        [
+                            'id' => 'active_off',
+                            'value' => false,
+                            'label' => $this->module->l('Disabled', self::FILE_NAME),
+                        ],
+                    ],
+                ];
+            }
 
-		AssortUtility::aasort($orderStatuses, 'id_order_state');
+            $isStatusAwaiting = Config::MOLLIE_AWAITING_PAYMENT === $status['name'];
+            $input[] = [
+                'type' => 'select',
+                'label' => $status['message'],
+                'tab' => $advancedSettings,
+                'desc' => $status['description'],
+                'name' => $status['key'],
+                'options' => [
+                    'query' => $isStatusAwaiting ? $allStatuses : $allStatusesWithSkipOption,
+                    'id' => 'id_order_state',
+                    'name' => 'name',
+                ],
+            ];
+        }
+        $input = array_merge($input, [
+            [
+                'type' => 'mollie-h2',
+                'name' => '',
+                'tab' => $advancedSettings,
+                'title' => $this->module->l('Visual Settings', self::FILE_NAME),
+            ],
+            [
+                'type' => 'select',
+                'label' => $this->module->l('Images', self::FILE_NAME),
+                'tab' => $advancedSettings,
+                'desc' => $this->module->l('Show big, normal or no payment method logos on checkout.', self::FILE_NAME),
+                'name' => Config::MOLLIE_IMAGES,
+                'options' => [
+                    'query' => [
+                        [
+                            'id' => Config::LOGOS_HIDE,
+                            'name' => $this->module->l('hide', self::FILE_NAME),
+                        ],
+                        [
+                            'id' => Config::LOGOS_NORMAL,
+                            'name' => $this->module->l('normal', self::FILE_NAME),
+                        ],
+                        [
+                            'id' => Config::LOGOS_BIG,
+                            'name' => $this->module->l('big', self::FILE_NAME),
+                        ],
+                    ],
+                    'id' => 'id',
+                    'name' => 'name',
+                ],
+            ],
+            [
+                'type' => 'text',
+                'label' => $this->module->l('CSS file', self::FILE_NAME),
+                'tab' => $advancedSettings,
+                'desc' => TagsUtility::ppTags(
+                    $this->module->l('Leave empty for default stylesheet. Should include file path when set. Hint: You can use [1]{BASE}[/1], [1]{THEME}[/1], [1]{CSS}[/1], [1]{MOBILE}[/1], [1]{MOBILE_CSS}[/1] and [1]{OVERRIDE}[/1] for easy folder mapping.', self::FILE_NAME),
+                    [$this->module->display($this->module->getPathUri(), 'views/templates/front/kbd.tpl')]
+                ),
+                'name' => Config::MOLLIE_CSS,
+                'class' => 'long-text',
+            ],
+        ]);
+        $input[] = [
+            'type' => 'mollie-carriers',
+            'label' => $this->module->l('Shipment information', self::FILE_NAME),
+            'tab' => $advancedSettings,
+            'name' => Config::MOLLIE_TRACKING_URLS,
+            'depends' => Config::MOLLIE_API,
+            'depends_value' => Config::MOLLIE_ORDERS_API,
+            'carriers' => $this->carrierInformationService->getAllCarriersInformation($this->lang->id),
+        ];
+        $input[] = [
+            'type' => 'mollie-carrier-switch',
+            'label' => $this->module->l('Automatically ship on marked statuses', self::FILE_NAME),
+            'tab' => $advancedSettings,
+            'name' => Config::MOLLIE_AUTO_SHIP_MAIN,
+            'desc' => $this->module->l('Enabling this feature will automatically send shipment information when an order gets marked status.', self::FILE_NAME),
+            'is_bool' => true,
+            'values' => [
+                [
+                    'id' => 'active_on',
+                    'value' => true,
+                    'label' => $this->module->l('Enabled', self::FILE_NAME),
+                ],
+                [
+                    'id' => 'active_off',
+                    'value' => false,
+                    'label' => $this->module->l('Disabled', self::FILE_NAME),
+                ],
+            ],
+            'depends' => Config::MOLLIE_API,
+            'depends_value' => Config::MOLLIE_ORDERS_API,
+        ];
+        $input[] = [
+            'type' => 'checkbox',
+            'label' => $this->module->l('Automatically ship when one of these statuses is reached', self::FILE_NAME),
+            'tab' => $advancedSettings,
+            'desc' => $this->module->l('If an order reaches one of these statuses the module will automatically send shipment information', self::FILE_NAME),
+            'name' => Config::MOLLIE_AUTO_SHIP_STATUSES,
+            'multiple' => true,
+            'values' => [
+                'query' => $orderStatuses,
+                'id' => 'id_order_state',
+                'name' => 'name',
+            ],
+            'expand' => (count($orderStatuses) > 10) ? [
+                'print_total' => count($orderStatuses),
+                'default' => 'show',
+                'show' => ['text' => $this->module->l('Show', self::FILE_NAME), 'icon' => 'plus-sign-alt'],
+                'hide' => ['text' => $this->module->l('Hide', self::FILE_NAME), 'icon' => 'minus-sign-alt'],
+            ] : null,
+            'depends' => Config::MOLLIE_API,
+            'depends_value' => Config::MOLLIE_ORDERS_API,
+        ];
+        $orderStatuses = [
+            [
+                'name' => $this->module->l('Disable this status', self::FILE_NAME),
+                'id_order_state' => '0',
+            ],
+        ];
+        $orderStatuses = array_merge($orderStatuses, OrderState::getOrderStates($this->lang->id));
+        $orderStatusesCount = count($orderStatuses);
+        for ($i = 0; $i < $orderStatusesCount; ++$i) {
+            $orderStatuses[$i]['name'] = $orderStatuses[$i]['id_order_state'] . ' - ' . $orderStatuses[$i]['name'];
+        }
 
-		$this->smarty->assign([
-			'logs' => $this->link->getAdminLink('AdminLogs'),
-		]);
-		$input = array_merge(
-			$input,
-			[
-				[
-					'type' => 'mollie-h2',
-					'name' => '',
-					'title' => $this->module->l('Debug level', self::FILE_NAME),
-					'tab' => $advancedSettings,
-				],
-				[
-					'type' => 'switch',
-					'label' => $this->module->l('Display errors', self::FILE_NAME),
-					'tab' => $advancedSettings,
-					'name' => Config::MOLLIE_DISPLAY_ERRORS,
-					'desc' => $this->module->l('Enabling this feature will display error messages (if any) on the front page. Use for debug purposes only!', self::FILE_NAME),
-					'is_bool' => true,
-					'values' => [
-						[
-							'id' => 'active_on',
-							'value' => true,
-							'label' => $this->module->l('Enabled', self::FILE_NAME),
-						],
-						[
-							'id' => 'active_off',
-							'value' => false,
-							'label' => $this->module->l('Disabled', self::FILE_NAME),
-						],
-					],
-				],
-				[
-					'type' => 'select',
-					'label' => $this->module->l('Log level', self::FILE_NAME),
-					'tab' => $advancedSettings,
-					'desc' => TagsUtility::ppTags(
-						$this->module->l('Recommended level: Errors. Set to Everything to monitor incoming webhook requests. [1]View logs.[/1]', self::FILE_NAME),
-						[
-							$this->module->display($this->module->getPathUri(), 'views/templates/admin/view_logs.tpl'),
-						]
-					),
-					'name' => Config::MOLLIE_DEBUG_LOG,
-					'options' => [
-						'query' => [
-							[
-								'id' => Config::DEBUG_LOG_NONE,
-								'name' => $this->module->l('Nothing', self::FILE_NAME),
-							],
-							[
-								'id' => Config::DEBUG_LOG_ERRORS,
-								'name' => $this->module->l('Errors', self::FILE_NAME),
-							],
-							[
-								'id' => Config::DEBUG_LOG_ALL,
-								'name' => $this->module->l('Everything', self::FILE_NAME),
-							],
-						],
-						'id' => 'id',
-						'name' => 'name',
-					],
-				],
-			]
-		);
+        AssortUtility::aasort($orderStatuses, 'id_order_state');
+
+        $this->smarty->assign([
+            'logs' => $this->link->getAdminLink('AdminLogs'),
+        ]);
+        $input = array_merge(
+            $input,
+            [
+                [
+                    'type' => 'mollie-h2',
+                    'name' => '',
+                    'title' => $this->module->l('Debug level', self::FILE_NAME),
+                    'tab' => $advancedSettings,
+                ],
+                [
+                    'type' => 'switch',
+                    'label' => $this->module->l('Display errors', self::FILE_NAME),
+                    'tab' => $advancedSettings,
+                    'name' => Config::MOLLIE_DISPLAY_ERRORS,
+                    'desc' => $this->module->l('Enabling this feature will display error messages (if any) on the front page. Use for debug purposes only!', self::FILE_NAME),
+                    'is_bool' => true,
+                    'values' => [
+                        [
+                            'id' => 'active_on',
+                            'value' => true,
+                            'label' => $this->module->l('Enabled', self::FILE_NAME),
+                        ],
+                        [
+                            'id' => 'active_off',
+                            'value' => false,
+                            'label' => $this->module->l('Disabled', self::FILE_NAME),
+                        ],
+                    ],
+                ],
+                [
+                    'type' => 'select',
+                    'label' => $this->module->l('Log level', self::FILE_NAME),
+                    'tab' => $advancedSettings,
+                    'desc' => TagsUtility::ppTags(
+                        $this->module->l('Recommended level: Errors. Set to Everything to monitor incoming webhook requests. [1]View logs.[/1]', self::FILE_NAME),
+                        [
+                            $this->module->display($this->module->getPathUri(), 'views/templates/admin/view_logs.tpl'),
+                        ]
+                    ),
+                    'name' => Config::MOLLIE_DEBUG_LOG,
+                    'options' => [
+                        'query' => [
+                            [
+                                'id' => Config::DEBUG_LOG_NONE,
+                                'name' => $this->module->l('Nothing', self::FILE_NAME),
+                            ],
+                            [
+                                'id' => Config::DEBUG_LOG_ERRORS,
+                                'name' => $this->module->l('Errors', self::FILE_NAME),
+                            ],
+                            [
+                                'id' => Config::DEBUG_LOG_ALL,
+                                'name' => $this->module->l('Everything', self::FILE_NAME),
+                            ],
+                        ],
+                        'id' => 'id',
+                        'name' => 'name',
+                    ],
+                ],
+            ]
+        );
 
         $fields = [
             'form' => [
@@ -822,6 +821,11 @@ class FormBuilder
             ],
         ];
 
+        return $fields;
+    }
+
+	public function buildAdvancedSettingsForm()
+	{
         $helper = new HelperForm();
 
         $helper->show_toolbar = false;
@@ -832,9 +836,7 @@ class FormBuilder
 
         $helper->identifier = $this->module->getIdentifier();
         $helper->submit_action = 'submitAdvancedSettings';
-        $helper->currentIndex = $this->module->getContext()->link->getAdminLink('AdminModules', false)
-            . "&configure={$this->module->name}&tab_module={$this->module->tab}&module_name={$this->module->name}";
-        $helper->token = Tools::getAdminTokenLite('AdminModules');
+        $helper->token = Tools::getAdminTokenLite(Mollie::ADMIN_MOLLIE_ADVANCED_SETTINGS_CONTROLLER);
 
         $helper->tpl_vars = [
             'fields_value' => $this->configFieldService->getConfigFieldsValues(),
@@ -842,6 +844,6 @@ class FormBuilder
             'id_language' => $this->module->getContext()->language->id,
         ];
 
-        return $helper->generateForm([$fields]);
+        return $helper->generateForm([$this->getAdvancedSettingsSection()]);
 	}
 }
